condensed object, component and system into ecs header
