+++ include/framework/systems/ControlSystem.hpp
@@ -1,15 +1,35 @@
 #ifndef HESTIA_FRAMEWORK_CONTROL_SYSTEM_HPP_
 #define HESTIA_FRAMEWORK_CONTROL_SYSTEM_HPP_
 
-#include "framework/ecs/System.hpp"
-#include "framework/ecs/ComponentManager.hpp"
+#include <map>
+
+#include "framework/ecs/ecs.h"
+
 #include "input/InputManager.hpp"
-#include "ControlComponent.hpp"
+#include "input/InputTypes.hpp"
 
 #include "util/Logger.hpp"
 
 namespace HGE {
 
+    /**
+     * Control Component
+     */
+    struct ControlComponent : public Component {
+
+        std::map<KeyType, bool> mKeys;
+
+        void addKey(KeyType key);
+        void removeKey(KeyType key);
+        bool getKeyValue(KeyType key);
+
+        ControlComponent(UID ownerId) : Component(ownerId), mKeys(std::map<KeyType, bool>()) { }
+        ~ControlComponent() { }
+    };
+
+    /**
+     * Control System
+     */
     template <>
     class System<ControlComponent> : public ISystem {
