--- src/framework/ecs/Object.cpp
@@ -1,10 +1,0 @@
-#include "framework/ecs/Object.hpp"
-
-#include "framework/Engine.hpp"
-
-namespace HGE {
-
-    // void Object::registerComponent(Component* component) {
-    //     //Engine::getInstance()->getComponentManager()->registerComponent(component);
-    // }
-}
