+++ src/framework/systems/ControlSystem.cpp
@@ -1,9 +1,32 @@
-#include "framework/control/ControlSystem.hpp"
-
+#include "framework/systems/ControlSystem.hpp"
 #include "framework/Engine.hpp"
 
 namespace HGE {
 
+    /**
+     * Control Component Methods
+     */    
+    void ControlComponent::addKey(KeyType key) {
+        mKeys[key] = false;
+    }
+
+    void ControlComponent::removeKey(KeyType key) {
+        auto it = mKeys.find(key);
+        mKeys.erase(it);
+    }
+
+    bool ControlComponent::getKeyValue(KeyType key) {
+        auto it = mKeys.find(key);
+        if(it != mKeys.end()) {
+            return mKeys[key];
+        } else {
+            return false;
+        }
+    }
+
+    /**
+     * Control System Methods
+     */    
     System<ControlComponent>::System(ComponentArray<ControlComponent>* componentArray) : mComponentsArray(componentArray) { 
         mInputManager = Engine::getInstance()->getInputManager();
         Logger::getInstance()->logDebug("Control System", "Created");
